Split up resolvers and typeDefs for organization. Also added pet and conversation typeDefs
